requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
aiohttp>=3.8.0
pyyaml>=6.0.0
statsmodels>=0.13.0
ta>=0.10.0
lxml>=4.8.0
beautifulsoup4>=4.10.0
html5lib>=1.1.0
python-dotenv>=0.19.0